requirements.txt
playwright==1.41.0
python-dotenv==1.0.0
pandas==2.1.4
google-generativeai==0.3.1
Pillow==10.1.0
